requirements.txt
aiohttp==3.11.11
asyncio==3.4.3
colorama==0.4.6
curl-cffi==0.7.4
loguru==0.7.0
python-dotenv==1.0.1
requests==2.32.3